Initial implementation of roving classifier
Uses classification accuracy as fitness. Does not seem to work very
well: it never classifies anything and still gets 90% fitness.

+++ edd/main.cpp
@@ -400,7 +400,7 @@ int main(int argc, char *argv[])
         
         eddAvgFitness /= (double)populationSize;
 		
-		cout << "generation " << update << ": edd [" << (int)eddAvgFitness << " : " << (int)eddMaxFitness << "]" << endl;
+		cout << "generation " << update << ": edd [" << eddAvgFitness << " : " << eddMaxFitness << "]" << endl;
         
         // display video of simulation
         if (make_interval_video)
